fix: fix buttons in header
