+++ src/components/Header/Header.tsx
@@ -1,10 +1,26 @@
-import React from 'react';
 import { Link } from 'react-router-dom';
 import styles from './Header.module.css';
 import logo from '../../assets/img/Logo.png';
 import langchange from '../../assets/img/Langchange.png';
+import { clearLocalStorage } from '../../utils/utils';
+import { useAuth } from '../hook/useAuth';
 
 function Header() {
+  const token = localStorage.getItem('token');
+  const [, dispatch] = useAuth();
+
+  function handleLogOut() {
+    clearLocalStorage();
+    dispatch({
+      type: 'user',
+      data: {
+        username: null,
+        login: null,
+        token: null,
+        id: null,
+      },
+    });
+  }
   return (
     <>
       <header>
@@ -14,14 +30,23 @@ function Header() {
           </div>
           <div className={styles.nav__wrapper}>
             <img className={styles.lang} src={langchange} alt="choose lang" />
-            <nav>
-              <Link to="/login">
-                <button className={styles.sign__in}>Sign In</button>
-              </Link>
-              <Link to="/register">
-                <button className={styles.sign__up}>Sign Up</button>
-              </Link>
-            </nav>
+            {!token && (
+              <nav>
+                <Link to="/login">
+                  <button className={styles.sign__in}>Sign In</button>
+                </Link>
+                <Link to="/register">
+                  <button className={styles.sign__up}>Sign Up</button>
+                </Link>
+              </nav>
+            )}
+            {token && (
+              <nav>
+                <button className={styles.sign__out} onClick={handleLogOut}>
+                  Sign Out
+                </button>
+              </nav>
+            )}
           </div>
         </div>
       </header>
